debundle/e2e: consolidate single-purpose test files into two domain-grouped ones
Many e2e files held a single test, all variations on "feed a
spec into the materializer and assert behaviour". Splitting one
test per file was a holdover from landing each as a distinct
bug-fix repro; the file boundaries no longer signal anything
meaningful and make it harder to scan related cases together.

Two domain-grouped files now hold the consolidated set:

- `realizability_test.rs` — 7 tests covering the validator's
  `I ∪ S` realizability gate. Sections: R cycles, I cycles via
  lazy back-edges, acyclic specs and cross-module init order,
  side-effect ordering edges (S), per-declarator owner
  attribution across comma-list var-decls. Replaces:
  `validator_rejects_cycle_test`, `lazy_back_edge_cycle_test`,
  `init_call_order_test`, `side_effect_ordering_test`,
  `comma_list_split_owners_test`.

- `cross_module_emission_test.rs` — 9 tests covering emit-side
  imports/exports/disambiguation. Sections: cross-module
  dependency wiring, consumer-side import-local
  disambiguation, ImportSpecifier-bound members, multi-module
  re-export of a shared imported binding, source-chunk
  re-imports for moved bodies. Replaces: `cross_module_test`,
  `import_collision_test`, `import_specifier_member_test`,
  `multi_module_re_export_test`, `runtime_import_reference_test`.

The AST-walking helpers from the old `import_collision_test`
(`parse_module`, `assert_unique_import_locals`,
`assert_export_named_specifier`,
`assert_unique_lexical_decls_per_scope`) move into
`support.rs` so they're reusable. The `:support` library gains
`@crates//:swc_common`, `@crates//:swc_ecma_ast`, and
`@crates//:swc_ecma_parser` deps.

`e2e/BUILD.bazel`: 18 `rust_test` targets → 10. Test count is
preserved (16 tests across the 10 deleted files; 16 tests in
the two new files); `bbr test //devinfra/js/debundle/...`
passes 14 targets.

Other e2e files unchanged: they each cover a distinct concern
with focused naming (`lowering_test`, `vendor_swap_test`,
`naturalization_test`, `scrambled_id_frequencies_test`,
`url_rebase_test`, `missing_binding_member_test`,
`sibling_declarator_steal_test`, `source_order_emit_test`).
